+++ InstabugSample/ios/InstabugSampleTests/InstabugSurveysTests.m
@@ -27,6 +27,8 @@
 
 @end
 
+// typedef void (^AvailableSurveysWithCompletionBlock)(NSArray<IBGSurveys *> *availableSurveys);
+
 @implementation InstabugSurveysTests
 
 - (void)setUp {
@@ -109,15 +111,15 @@
   [self waitForExpectationsWithTimeout:EXPECTATION_TIMEOUT handler:nil];
 }
 
-- (void) testGetAvailableSurveys {
-  id mock = OCMClassMock([IBGSurveys class]);
-  RCTResponseSenderBlock callback = ^void(NSArray *response) {};
-  
-  OCMStub([mock availableSurveys]);
-  [self.instabugBridge getAvailableSurveys:callback];
-  OCMVerify([mock availableSurveys]);
-  
-}
+// - (void) testGetAvailableSurveys {
+//   id mock = OCMClassMock([IBGSurveys class]);
+//   RCTResponseSenderBlock callback = ^void(NSArray *response) {};
+//   AvailableSurveysWithCompletionBlock deeperCallback = ^(NSArray<IBGSurveys *> *availableSurveys) {};
+
+//   OCMStub([mock availableSurveysWithCompletionHandler:deeperCallback]);
+//   [self.instabugBridge getAvailableSurveys:callback];
+//   OCMVerify([mock availableSurveysWithCompletionHandler:deeperCallback]);
+// }
 
 - (void) testSetWillShowSurveyHandler {
   id partialMock = OCMPartialMock(self.instabugBridge);
@@ -137,6 +139,14 @@
   OCMStub([partialMock sendEventWithName:OCMOCK_ANY body:nil]);
   IBGSurveys.didDismissSurveyHandler();
   OCMVerify([partialMock sendEventWithName:@"IBGDidDismissSurvey" body:nil]);
+}
+
+
+- (void) testSetAppStoreURL {
+  NSString *appStoreURL = @"http://test";
+  
+  [self.instabugBridge setAppStoreURL:appStoreURL];
+  XCTAssertEqual(IBGSurveys.appStoreURL, appStoreURL);
 }