+++ android/src/main/java/com/instabug/reactlibrary/RNInstabugReactnativeModule.java
@@ -159,6 +159,7 @@ public class RNInstabugReactnativeModule extends ReactContextBaseJavaModule {
 
     private final String INVOCATION_HEADER = "invocationHeader";
     private final String START_CHATS = "startChats";
+    private final String REPORT_QUESTION = "reportQuestion";
     private final String REPORT_BUG = "reportBug";
     private final String REPORT_FEEDBACK = "reportFeedback";
 
@@ -2229,7 +2230,11 @@ public class RNInstabugReactnativeModule extends ReactContextBaseJavaModule {
                         final View[] arrayOfViews = new View[ids.size()];
                         for (int i = 0; i < ids.size(); i++) {
                             int viewId = (int) ids.getDouble(i);
-                            arrayOfViews[i] = nativeViewHierarchyManager.resolveView(viewId);
+                            try {
+                                arrayOfViews[i] = nativeViewHierarchyManager.resolveView(viewId);
+                            } catch(Exception e) {
+                                e.printStackTrace();
+                            }
                         }
                         Instabug.setViewsAsPrivate(arrayOfViews);
                     }
@@ -2283,6 +2288,8 @@ public class RNInstabugReactnativeModule extends ReactContextBaseJavaModule {
                 return InstabugCustomTextPlaceHolder.Key.INVOCATION_HEADER;
             case START_CHATS:
                 return InstabugCustomTextPlaceHolder.Key.START_CHATS;
+            case REPORT_QUESTION:
+                return InstabugCustomTextPlaceHolder.Key.REPORT_QUESTION;
             case REPORT_BUG:
                 return InstabugCustomTextPlaceHolder.Key.REPORT_BUG;
             case REPORT_FEEDBACK:
@@ -2473,6 +2480,7 @@ public class RNInstabugReactnativeModule extends ReactContextBaseJavaModule {
         constants.put("commentFieldHintForQuestion", COMMENT_FIELD_HINT_FOR_QUESTION);
         constants.put("invocationHeader", INVOCATION_HEADER);
         constants.put("startChats", START_CHATS);
+        constants.put("reportQuestion", REPORT_QUESTION);
         constants.put("reportBug", REPORT_BUG);
         constants.put("reportFeedback", REPORT_FEEDBACK);
         constants.put("conversationsHeaderTitle", CONVERSATIONS_LIST_TITLE);