+++ index.js
@@ -158,6 +158,16 @@ const InstabugModule = {
     Instabug.setSessionProfilerEnabled(sessionProfilerEnabled);
   },
 
+  /**
+   * This API sets the verbosity level of logs used to debug The SDK. The defualt value in debug 
+   * mode is sdkDebugLogsLevelVerbose and in production is sdkDebugLogsLevelError.
+   * @param {sdkDebugLogsLevel} sdkDebugLogsLevel - The verbosity level of logs.
+   *
+   */
+  setSdkDebugLogsLevel(sdkDebugLogsLevel) {
+    Instabug.setSdkDebugLogsLevel(sdkDebugLogsLevel);
+  },
+
   /* istanbul ignore next */
   /**
    * @deprecated use {@link Replies.getUnreadRepliesCount}
@@ -858,6 +868,19 @@ const InstabugModule = {
   },
 
   /**
+   * Verbosity level of the SDK debug logs. This has nothing to do with IBGLog,
+   * and only affect the logs used to debug the SDK itself.
+   * @readonly
+   * @enum {number}
+   */
+  sdkDebugLogsLevel: {
+    sdkDebugLogsLevelVerbose: Instabug.sdkDebugLogsLevelVerbose,
+    sdkDebugLogsLevelDebug: Instabug.sdkDebugLogsLevelDebug,
+    sdkDebugLogsLevelError: Instabug.sdkDebugLogsLevelError,
+    sdkDebugLogsLevelNone: Instabug.sdkDebugLogsLevelNone,
+  },
+
+  /**
    *  The extended bug report mode
    * @readonly
    * @enum {number}
@@ -965,6 +988,7 @@ const InstabugModule = {
     invalidCommentTitle: Instabug.invalidCommentTitle,
     invocationHeader: Instabug.invocationHeader,
     startChats: Instabug.startChats,
+    reportQuestion: Instabug.reportQuestion,
     reportBug: Instabug.reportBug,
     reportFeedback: Instabug.reportFeedback,
     emailFieldHint: Instabug.emailFieldHint,